initial packaging of both gobby and infinote
infinote is not yet stable API- and ABI-wise so I chose to link it
statically into gobby 0.5, the version based on it.  Currently
this does not fully work yet because there are two plugins
required by Gobby and the dedicated server infinoted.  Those will
be eliminated for the time being, so that no infinote library
package needs to be introduced.

Signed-off-by: Philipp Kern <pkern@debian.org>